V2 Board Updates
Added in extra resistors, cap, and diode that were added to a  particular version of the SI board from the USDM on the comparator circuit. Updated the schematic diagram of the comparator circuit to improve readability. Updated the clock frequency measured with a scope. Made a few minor layout improvements including moving the "JP" connections inwards as not to interfere with the plastic housing for the PCB in the cluster and other minor layout errors such as swapped reference designators. Updated info on where the +5V regulator gets its source voltage from. These updates are reflected in schematic version V2.1 as denoted in the title box (not in the title of the schematic file).